refactor: some structure minor refactor
