+++ src/const/const.ts
@@ -1,0 +1,3 @@
+export const FILE_SIZE = 1024 * 1024;
+export const SUPPORTED_FORMATS = ["image/jpg", "image/jpeg", "image/png"];
+export const GENDER = ["male", "female", "other"];
