+++ src/routes/uncontrolled-form/UncontrolledForm.tsx
@@ -1,97 +1,16 @@
-import { countries } from "countries-list";
 import { useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import * as yup from "yup";
 
 import { addFormData } from "../../app/formDataSlice";
-import { fileToBase64 } from "../../app/helpers/fileToBase64";
 import { useAppDispatch, useAppSelector } from "../../app/hooks";
 import { RootState } from "../../app/store";
 import Header from "../../components/Header";
 import PasswordStrengthMeter from "../../components/PasswordStrengthMeter";
+import { GENDER } from "../../const/const";
+import { fileToBase64 } from "../../helpers/fileToBase64";
 import { UncontrolledFormInput } from "../../types/types";
-
-const FILE_SIZE = 1024 * 1024;
-const SUPPORTED_FORMATS = ["image/jpg", "image/jpeg", "image/png"];
-const GENDER = ["male", "female", "other"];
-
-const firstUppercaseLetterSchema = yup
-  .string()
-  .required("name is required")
-  .test(
-    "first-uppercase-letter",
-    "the first letter must be uppercase",
-    (value) => {
-      if (!value) return false;
-      return /^[A-Z]/.test(value);
-    },
-  );
-
-const passwordSchema = yup
-  .string()
-  .required("password is required")
-  .matches(/[a-z]/, "password must contain at least one lowercase letter")
-  .matches(/[A-Z]/, "password must contain at least one uppercase letter")
-  .matches(/\d/, "password must contain at least one number")
-  .matches(
-    /[~`!@#$%^&*()_\-+={[}\]|\\:;"'<,>.?/]/,
-    "password must contain at least one special character",
-  )
-  .min(8, "password must be at least 8 characters long");
-
-const schema = yup.object().shape({
-  name: firstUppercaseLetterSchema,
-  age: yup
-    .number()
-    .required("age is required")
-    .typeError("age must be a number")
-    .max(100, "age must be less than 100")
-    .positive()
-    .integer(),
-  email: yup.string().required("email is required").email(),
-  password: passwordSchema,
-  confirmPassword: yup
-    .string()
-    .required("confirm Password is required")
-    .oneOf([yup.ref("password")], "passwords must match"),
-  gender: yup
-    .mixed<"male" | "female" | "other">()
-    .required("gender is required")
-    .oneOf(["male", "female", "other"]),
-  acceptTermsConditions: yup
-    .boolean()
-    .required("accept T&C is required")
-    .oneOf([true], "accept T&C is required"),
-  picture: yup
-    .mixed<FileList>()
-    .required("file is required")
-    .test("required", "file is required", (files) => {
-      return files && files.length > 0;
-    })
-    .test(
-      "fileSize",
-      `file size must not be more than ${FILE_SIZE}`,
-      (files) => {
-        if (files[0]) {
-          return files[0].size <= FILE_SIZE;
-        }
-        return true;
-      },
-    )
-    .test("fileFormat", "only jpeg and png files are allowed", (files) => {
-      if (files[0]) {
-        return SUPPORTED_FORMATS.includes(files[0].type);
-      }
-      return true;
-    }),
-  country: yup
-    .string()
-    .required("country is required")
-    .oneOf(
-      Object.values(countries).map((country) => country.name),
-      "invalid country",
-    ),
-});
+import { schema } from "../../validation/validation";
 
 export default function UnControlledForm() {
   const [errors, setErrors] = useState<{ [key: string]: { message: string } }>(
@@ -144,7 +63,6 @@ export default function UnControlledForm() {
       picture: pictureRef.current?.files,
       country: countryRef.current?.value,
     };
-    console.log(data);
 
     if (!(await validate(data))) {
       return;
